+++ src/world/import/SceneImporter.cpp
@@ -1,0 +1,7 @@
+#include "world/import/SceneImporter.h"
+
+namespace world {
+
+  SceneImporter::~SceneImporter() = default;
+
+}